refactor: limpiar estructura de menus, agregar funcionalidad faltante a menu_usuario_estandar
